wifi_hal: Remove "invalid log_type" logged on an unknown pktlog type
All the unparsed pktlog V2 event types are logged to help analyze
such types being received. Such logs are ended up to be huge and
are thus results in heavy mips consumption for event loop.
These ended up in this thread not even receiving the stop indication
on the exit pair sockets and thus the WiFi HAL Stop fails.

This further was leading to Wi-Fi ON failures, for the failed Wi-Fi HAL
stop/OFF.

Hence, removing such spurious logs

CRs-Fixed: 2555808
Change-Id: Icd1adc3de6474c60d8b4745d8b4b3c13c7646d7c